requirements.txt
gymnasium[box2d]==0.28.1
box2d-py==2.3.5
torch==2.1.0
numpy==1.24.3
matplotlib==3.7.2
pyyaml==6.0.1
tensorboard==2.14.0
tqdm==4.66.1
seaborn==0.12.2
opencv-python==4.8.1.78
imageio==2.31.5
imageio-ffmpeg==0.4.9 